Refactor layout, fix vertical centering in some situations
Use flexbox layout and image object-fit property to layout all the different modes. This should be a bit clearer, and we no longer need to script any particular margin lengths anymore, layouts are fully handled by CSS classes.

This fixes a few situations where the image was not always vertically centered as intended.

Curiously, when trying to use the flexbox properties:
```
align-items: center;
justify-content: center;
```
they just seem to have no effect in this particular situation. Instead we use `margin: auto` on the flex item, which works as intended.